+++ .gitignore
@@ -30,3 +30,5 @@
 
 # Debug files
 *.dSYM/
+
+rocket_launch
