+++ README.md
@@ -1,2 +1,9 @@
 # my-printf
 Rewriting printf
+
+# Build
+make && make clean
+
+# TODO
+- [x] Build arbo
+- [ ] Package
