+++ include/puts/puts_char.h
@@ -1,0 +1,13 @@
+/******************************************************************************
+ *
+ * File Name        : my_printf.h
+ * Created By       : Hugo POULIQUEN
+ * Creation Date    : 01 07th, 2016
+ * Last Changed By  : Hugo POULIQUEN
+ * Last Change      : 03 13th, 2016
+ * Purpose          : Description
+ *
+ *******************************************************************************/
+
+char my_puts(char str[]);
+char my_puts_char(char *str);
