+++ include/puts/puts_nbr.h
@@ -1,0 +1,12 @@
+/******************************************************************************
+ *
+ * File Name        : puts_nbr.h
+ * Created By       : Hugo POULIQUEN
+ * Creation Date    : 01 07th, 2016
+ * Last Changed By  : Hugo POULIQUEN
+ * Last Change      : 03 13th, 2016
+ * Purpose          : Description
+ *
+ *******************************************************************************/
+
+ int my_put_nbr(int nbr);
