+++ include/str/str_cat.h
@@ -1,0 +1,13 @@
+/******************************************************************************
+ *
+ * File Name        : my_printf.h
+ * Created By       : Hugo POULIQUEN
+ * Creation Date    : 01 07th, 2016
+ * Last Changed By  : Hugo POULIQUEN
+ * Last Change      : 03 13th, 2016
+ * Purpose          : Description
+ *
+ *******************************************************************************/
+
+char my_strcat(char *src, char *dest);
+char my_strncat(char *src, char *dest, int n);
