+++ str/str_comp.c
