+++ str/str_copy.c
