+++ tests/test.c
@@ -1,0 +1,5 @@
+#include <criterion/criterion.h>
+
+Test(simple, test){
+    cr_assert(0, "Hello world !");
+}
